chore: move session specific tests from test_client to test_session
